requirements.txt
argparse
easydict
h5py
matplotlib
numpy
open3d==0.14.1
opencv-python
pyyaml
scipy
tensorboardX
timm==0.4.5
tqdm
transforms3d
pytorch3d
